pyproject.toml
[tool.poetry]
name = "tusk-api"
version = "0.1.0"
description = "The drf api for tusk app"
authors = ["Mikhail Vialov <mishavjalv48@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tusk_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.5"
django-types = "^0.16.0"
django-filter = "^22.1"

[tool.poetry.group.dev.dependencies]
pynvim = "^0.4.3"
black = "^22.10.0"
flake8 = "^6.0.0"
django-environ = "^0.9.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
drf-spectacular = "^0.24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
